+++ src/main/java/com/example/todoshpp/model/attribut/Status.java
@@ -5,4 +5,5 @@ public enum Status {
     WORK_IN_PROGRESS,
     DONE,
     CANCELLED;
+
 }
